feat: Enhance deployment file clearing script with help argument handling

+++ packages/snfoundry/deployments/clear.mjs
@@ -4,21 +4,23 @@ import path from "path";
 const k = ["devnet", "sepolia", "mainnet"];
 const args = process.argv.slice(2);
 
-const match = k.some((x) => args[1]?.includes(x) ?? false);
+if (args.some((x) => !x.includes("help"))) {
+  const match = k.some((x) => args[1]?.includes(x) ?? false);
 
-// Deletes every temporary file found.
-fs.readdir(".", (e, f) => {
-  if (e) console.error("Error reading directory: ", e);
+  // Deletes every temporary file found.
+  fs.readdir(".", (e, f) => {
+    if (e) console.error("Error reading directory: ", e);
 
-  let mf;
-  if (!match) mf = f.filter((x) => k.some((w) => x.startsWith(w)));
-  else mf = f.filter((x) => x.startsWith(args[0].substring(2)));
+    let mf;
+    if (!match) mf = f.filter((x) => k.some((w) => x.startsWith(w)));
+    else mf = f.filter((x) => x.startsWith(args[0].substring(2)));
 
-  for (let i = 0; i < mf.length; i++) {
-    const f = path.join(".", mf[i]);
+    for (let i = 0; i < mf.length; i++) {
+      const f = path.join(".", mf[i]);
 
-    fs.unlink(f, (e) => {});
-  }
+      fs.unlink(f, (e) => {});
+    }
 
-  console.log("\x1b[32m", "temporary files deleted.");
-});
+    console.log("\x1b[32m", "temporary files deleted.");
+  });
+}
